% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coxdata}
\alias{coxdata}
\title{Cox Modeling Dataset}
\format{
A data frame with 194 rows and 6 variables:
\describe{
\item{TimeToEvent}{Numeric: time to event}
\item{Outcome}{Numeric: event indicator (1=event, 0=censored)}
\item{C.Index}{Numeric: C-index grouping variable}
\item{BNP}{Numeric: brain natriuretic peptide}
\item{LVEF}{Numeric: left ventricular ejection fraction}
\item{gender}{Factor: subject gender (F/M)}
}
}
\usage{
coxdata
}
\description{
Time-to-event data from 194 subjects for Cox proportional hazards modeling.
}
\keyword{datasets}
