% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu1way.R
\name{cu1way}
\alias{cu1way}
\title{Does one-way anova/post-hoc tests or contingency tables, and makes a bar graph}
\usage{
cu1way(depvar, group1, ebars = 0, ordinal = NULL, plot = "bar",
  ytrans = "none", scale = "frequency", dots = 0, barcolor = "black",
  barfill = "colors", casecontrol = F, maxfor2 = 5, g1order = NULL,
  psigcld = 0, conf.int = 0.95, refmean = NA, depname = NULL,
  g1name = NULL, title = NULL, caption = NULL, minimal = F,
  padj = "none", pbart = 0.01, pnorm = 0.01, chariqr = ":",
  pvpairs = "std", pvypos = NULL, pvstinc = 0.05, pvlab = "p",
  pvprefix = "p=", pvsize = NULL, chpvref = "ref", pvspill = F,
  pnosig = 0.2, psignif = 0.05, p2stars = 0.01, p3stars = 0.001,
  p4stars = 1e-04, pvnshide = T, pvtipl = 0.01, doAll = T,
  linetype = "n", linesize = 1, linecolor = "red", theme = "bw",
  legend = "top", size = NULL, width = NULL, yscale = "none",
  letleft = T, nletbarmax = 3, font.x = NULL, xangle = NULL,
  yangle = NULL, orientation = "vertical", dotsize = NULL,
  dotshape = NULL, dotcolor = NULL, posd = NULL, binwfac = 30,
  fontfamily = "sans", ymin = NA, ymax = NA, fontmain = NULL,
  fontxname = NULL, fontyname = NULL, fontxticks = NULL,
  fontyticks = NULL, axiscolor = NULL, tickcolor = NULL,
  axisthick = NULL, tickthick = NULL, ticklength = NULL,
  xticks.by = NULL, yticks.by = NULL, titlejust = "center",
  legheadsize = 12, legtextsize = 10, ftype = NULL, fname = NULL,
  fscale = NULL, fwidth = NULL, fheight = NULL, dpi = 300,
  remove = NULL)
}
\arguments{
\item{depvar}{required: variable being analyzed}

\item{group1}{required: grouping factor}

\item{ebars}{=0 (default to be 1 or 4)/1/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR) or -N (nonparm if any norm fail)}

\item{ordinal}{=NULL (default)/T/c("...") to treat variable as no/yes ordinal}

\item{plot}{="bar" (default) for bar graphs; "box" "violin" "rod" "no"}

\item{ytrans}{="none" (default) to transform depvar: "sqrt" "log"}

\item{scale}{="frequency" (default) or "percent"}

\item{dots}{=0 (default), 1 to display data on graph}

\item{barcolor}{="black" (default) for bar outline color}

\item{barfill}{="colors" (default) for bar fill color. Use fill="grey" ="black" etc for single color.}

\item{casecontrol}{=F (default)/T if depvar has 2 levels control and case}

\item{maxfor2}{=5 (default)/x for max p-values on a line for categorical data}

\item{g1order}{=NULL (default)/c("...") to reorder bars}

\item{psigcld}{=0 (no letters)/x for CLD letters on bars (any shared letter means P>psigcld, 0 for no letters)}

\item{conf.int}{=0.95 (default)/x/0 for confidence interval width of contrast estimates (0 for none)}

\item{refmean}{=NA (default)/ref.val to compare each group mean with}

\item{depname}{=NULL to override name of depvar}

\item{g1name}{=NULL to override name of group1}

\item{title}{=NULL to override plot title}

\item{caption}{=NULL (default)/"yes"/"caption text" to get caption ("yes" to list n's)}

\item{minimal}{=F (default)/T for minimal console output and no graph}

\item{padj}{="none"(default), p-value adjustment for multiple comparisons, can be "bonferroni", "holm")}

\item{pbart}{=0.05 (default)/x for Bartlett test threshold for variance homogeneity}

\item{pnorm}{=0.05 (default)/x for normality test threshold}

\item{chariqr}{=":" (default)/x for character to separate quartiles ("-" "," etc)}

\item{pvpairs}{="std" (default)/"all"/i/c(ij,ik,...) to show std or all or vs.i or pval's of i/j, i/k ...}

\item{pvypos}{=NULL (default)/position of pval lines}

\item{pvstinc}{=0.05 (default)/increment for pvypos from one pval line to next}

\item{pvlab}{="p" (default)/"*" to display numerical or asterisks}

\item{pvprefix}{="p=" (default)/"" etc prefix to p-values}

\item{pvsize}{=NULL (default 3.5, 7)/size of p-values or **}

\item{chpvref}{="ref" (default) char above i-th group bar when pvpairs=i}

\item{pvspill}{=F (default)/T to not allow p-values to spill outside range or to allow}

\item{pnosig}{=0.2 (default)/x for threshold to show p-values even if nonsigificant}

\item{psignif}{=0.05 (default)/x for threshold to significance}

\item{p2stars}{=0.01 (default)/x for threshold to two stars}

\item{p3stars}{=0.001 (default)/x for threshold to three stars}

\item{p4stars}{=0.0001 (default)/x for threshold to four stars}

\item{pvnshide}{=T (default)/F to hide NS p-values or not}

\item{pvtipl}{=0.01 (default)/length of p-value line tips}

\item{doAll}{=T (default) for "All" column; F for no; "I" in quotes for All first.}

\item{linetype}{="n" (default)/x for no connecting lines ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linesize}{=1 (default)/x for line thickness}

\item{linecolor}{="black" (default)/x for black lines ("red" etc)}

\item{theme}{="bw" (default)/x for white background (or "gray" or "classic" (no grid lines))}

\item{legend}{(default="top"), can be "bottom", "right", "left"}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{letleft}{=T (default) or F to letter the bars from the left (T) or from most similar (F)}

\item{nletbarmax}{=3 (default)/x for max consecutive letters on bar before using "-"}

\item{font.x}{=NULL (default)}

\item{xangle}{=NULL for x-axis value angle: 0 (default) horizontal, 90 vertical, or any value between}

\item{yangle}{=NULL for y-axis value angle: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can be "horizontal" or "reverse"}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{dotshape}{=NULL (default)}

\item{dotcolor}{=NULL (default "white") set to dot color}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=30 (default) set to fraction of range within which points will be binned}

\item{fontfamily}{="sans" (default), can be "serif" "mono"}

\item{ymin}{=NA (default) or value to start y-axis}

\item{ymax}{=NA (default) or value to end y-axis}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for no title}

\item{fontxname}{=c(12,"plain","black") (default) for x-axis name font, 0 to suppress}

\item{fontyname}{=c(12,"plain","black") (default) for y-axis name font, 0 to suppress}

\item{fontxticks}{=c(12,"plain","black") (default) for x-axis tick label font, 0 to suppress}

\item{fontyticks}{=c(12,"plain","black") (default) for y-axis tick label font, 0 to suppress}

\item{axiscolor}{="black" (default)/x for axis color}

\item{tickcolor}{="black" (default)/x for tick color}

\item{axisthick}{=0.5 (default)/x for axis thickness}

\item{tickthick}{=0.5 (default)/x for tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{=NULL (default)/s for x tick spacing by s}

\item{yticks.by}{=NULL (default)/s for y tick spacing by s}

\item{titlejust}{="center" (default) or "left" or "right"}

\item{legheadsize}{=12 (default) for font size of legend heading}

\item{legtextsize}{=10 (default) for the font size of legend text}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{=NULL(default) or set to numerical value for file scale}

\item{fwidth}{=NULL(default) or set to file width}

\item{fheight}{=NULL(default) or set to file height}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Does one-way anova/post-hoc tests or contingency tables, and makes a bar graph
}
\examples{
attach(NEJM)
cu1way(tcchange, Diet)  # for one-way anova and post-hoc t-tests, graphing mean+/-sd
# cu1way(tcchange, Diet, ebars=N)  # N=0 for 1 or 4, 1 for SD, 2 for SE, 3 for CL
cu1way(tcchange, Diet, ebars=4)  # for nonparametric Dunn
detach(NEJM)
attach(Met)
cu1way(feel, WTCAT) # contingency tables (all possible 2x2s)
detach(Met)
}
