% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu2way.R
\name{cu2way}
\alias{cu2way}
\title{Does two-way anova and post-hoc tests or contingency tables, and makes a bar graph}
\usage{
cu2way(depvar, group1, group2, interact = TRUE, dosimpler = F,
  partialF = TRUE, ordinal = NULL, scale = "frequency", ebars = 1,
  dots = 0, plot = "bar", ytrans = "none", barcolor = "black",
  barfill = "colors", casecontrol = F, maxfor2 = 5, g1order = NULL,
  g2order = NULL, psigcld = 0, conf.int = 0.95, depname = NULL,
  g1name = NULL, g2name = NULL, title = NULL, caption = NULL,
  minimal = F, padj = "none", pbart = 0.05, pnorm = 0.05,
  chariqr = ":", charamp = "&", pvpairs = "std", pvypos = NULL,
  pvstinc = 0.05, pvlab = "p", pvprefix = "p=", pvsize = NULL,
  chpvref = "ref", pvspill = F, pnosig = 0.2, psignif = 0.05,
  p2stars = 0.01, p3stars = 0.001, p4stars = 1e-04, pvnshide = T,
  pvtipl = 0.01, doAll = T, linetype = "n", linesize = 1,
  linecolor = "red", legend = "top", size = NULL, width = NULL,
  yscale = "none", letleft = T, nletbarmax = 3, xangle = NULL,
  yangle = NULL, orientation = "vertical", posd = NULL, binwfac = 30,
  dotsize = NULL, dotshape = NULL, theme = "bw", fontfamily = "sans",
  ymin = NA, ymax = NA, fontmain = NULL, fontxname = NULL,
  fontyname = NULL, fontxticks = NULL, fontyticks = NULL,
  axiscolor = NULL, tickcolor = NULL, axisthick = NULL,
  tickthick = NULL, ticklength = NULL, xticks.by = NULL,
  yticks.by = NULL, titlejust = "center", legheadsize = 12,
  legtextsize = 10, ftype = NULL, fname = NULL, fscale = NULL,
  fwidth = NULL, fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{depvar}{,group1,group2 required: variable being analyzed, grouping factors 1 & 2}

\item{interact}{=TRUE (default), F or FALSE for no interaction}

\item{dosimpler}{=F (default) or T not to do or to do 2 1ways first}

\item{partialF}{=TRUE (default), F or FALSE for no partial F vs simpler models}

\item{ordinal}{=NULL (default)/T/c("...") to treat variable as no/yes ordinal}

\item{scale}{="frequency" (default) or "percent"}

\item{ebars}{=1 (default)/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR) or -N (nonparm if any norm fail)}

\item{dots}{=0 (default), 1 to display data on graph}

\item{plot}{="bar" (default) for bar graphs; "box" "violin" "rod" "no"}

\item{ytrans}{="none" (default) to transform depvar: "sqrt" "log"}

\item{barfill}{="lancet" (default) for colors by group ("lancet" for 2 factors, other journal options "aaas", "jco", "uchicago", "npg"}

\item{casecontrol}{=F (default)/T if depvar has 2 levels control and case}

\item{maxfor2}{=5 (default)/x for max p-values on a line for categorical data}

\item{g1order}{=NULL (default) to reorder bars (1st Factor)}

\item{g2order}{=NULL (default) to reorder bars (2nd Factor)}

\item{psigcld}{=0 (no letters)/x for CLD letters on bars (any shared letter means P>psigcld, 0 for no letters)}

\item{conf.int}{=0.95 (default)/x/0 for confidence interval width of contrast estimates (0 for none)}

\item{depname}{/g1name/g2name/title=NULL to override names of dependent/Factor1/Factor2/title}

\item{caption}{=NULL (default)/"yes"/"caption text" to get caption ("yes" to list n's)}

\item{minimal}{=F (default)/T for minimal console output and no graph}

\item{padj}{="none"(default), p-value adjustment for multiple comparisons, can be "bonferroni", "holm")}

\item{pbart}{=0.05 (default)/x for Bartlett test threshold for variance homogeneity}

\item{pnorm}{=0.05 (default)/x for normality test threshold}

\item{chariqr}{=":" (default)/x for character to separate quartiles ("-" "," etc)}

\item{charamp}{="&" (default)/x for character to separate group1&2 levels ("-" etc)}

\item{pnosig}{=0.2 (default)/x for threshold to show p-values even if nonsigificant}

\item{psignif}{=0.05 (default)/x for threshold to significance}

\item{p2stars}{=0.01 (default)/x for threshold to two stars}

\item{p3stars}{=0.001 (default)/x for threshold to three stars}

\item{doAll}{=T (default) for "All" column; F for no; "I" in quotes for All first.}

\item{linetype}{="n" (default)/x for no connecting lines ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linesize}{=1 (default)/x for line thickness}

\item{legend}{(default="top"), can be "bottom", "right", "left"}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{letleft}{=T (default) or F to letter the bars from the left (T) or from most similar (F)}

\item{nletbarmax}{=3 (default)/x for max consecutive letters on bar before using "-"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can be "horizontal" or "reverse"}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=30 (default) set to fraction of range within which points will be binned}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{theme}{="bw" (default)/x for white background ("classic" (no grid lines),"linedraw" "gray" "minimal" "void")}

\item{fontfamily}{="sans" (default), can be "serif" "mono"}

\item{ymin, }{ymax =NA (default) or value to start/end y-axis}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for not title}

\item{fontxname}{.fontyname,fontxticks,fontyticks = c(12,"plain","black") default, 0 to suppress}

\item{axiscolor}{,tickcolor="black" (default)/x for axis/tick color}

\item{axisthick}{,tickthick=0.5 (default)/x for axis/tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{,yticks.by =NULL (default)/s for x/y tick spacing by s}

\item{titlejust}{="center" (default) or "left" or "right"}

\item{legheadsize}{=12 (default) for font size of legend heading}

\item{legtextsize}{=10 (default) for the font size of legend text}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}

\item{pvpairs="std"}{(default)/"all"/i/c(ij,ik,...) to show std or all or vs.i or pval's of i/j, i/k ...}

\item{pvypos=NULL}{(default)/position of pval lines}

\item{pvstinc=0.05}{(default)/increment for pvypos from one pval line to next}

\item{pvlab="p"/"*"}{to display numerical or asterisks}

\item{pvprefix="p="}{(default)/"" etc prefix to p-values}

\item{pvsize=NULL}{(default 3.5,7)/size of p-values or **}

\item{chpvref="ref"}{(default) char above i-th group bar when pvpairs=i}

\item{pvspill=F/T}{to not allow p-values to spill outside range or to allow}

\item{pvnshide=T}{(default)/F to hide NS p-values or not}

\item{pvtipl=0.01}{(default)/length of p-value line tips}
}
\value{
returns nothing
}
\description{
Does two-way anova and post-hoc tests or contingency tables, and makes a bar graph
}
\examples{
\dontrun{
cu2way(tcchange, Diet, sex)  # two-factor anova with interaction (all possible contrasts)
cu2way(tcchange, Diet, sex, interact=F)  # two-factor anova with no interaction (all possible contrasts)
cu2way(feel,WTCAT,Sex)  # contingency tables (all possible 2x2s)
}
}
