% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_2x2out.R
\name{cu_2x2out}
\alias{cu_2x2out}
\title{internal cufunction to print 2x2 and derived quantities}
\usage{
cu_2x2out(mytab, row1name, row2name)
}
\description{
internal cufunction to print 2x2 and derived quantities
}
\keyword{internal}
