% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_chi2fish.R
\name{cu_chi2fish}
\alias{cu_chi2fish}
\title{internal cufunction to calculate chi-squared and Fisher exact p-values
returns 6-list with chi-squared and Fisher p-values, RR/OR, conf.limits, SE}
\usage{
cu_chi2fish(contab, prelabel, casecontrol = F)
}
\description{
internal cufunction to calculate chi-squared and Fisher exact p-values
returns 6-list with chi-squared and Fisher p-values, RR/OR, conf.limits, SE
}
\keyword{internal}
