% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_chifishout.R
\name{cu_chifishout}
\alias{cu_chifishout}
\title{internal cufunction to display chi-squared and Fisher exact tables}
\usage{
cu_chifishout(
  casecont,
  rrortab,
  chitab,
  fishtab,
  nlev,
  namegroup,
  groupvar,
  maxfor2
)
}
\description{
internal cufunction to display chi-squared and Fisher exact tables
}
\keyword{internal}
