% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_dredge.R
\name{cu_dredge}
\alias{cu_dredge}
\title{internal cufunction to do dredge work, called by culogist, cucox, culinreg}
\usage{
cu_dredge(regobj, forlag, nmodshow = 16, m.min = 1, m.max = 99,
  fixed = NULL, subset = as.expression(1))
}
\description{
internal cufunction to do dredge work, called by culogist, cucox, culinreg
}
\keyword{internal}
