% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_estout.R
\name{cu_estout}
\alias{cu_estout}
\title{internal cufunction to print estimable results, return p-value}
\usage{
cu_estout(fit, vec, conf.int = 0.95, minimal = F)
}
\description{
internal cufunction to print estimable results, return p-value
}
\keyword{internal}
