% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_index.R
\name{cu_index}
\alias{cu_index}
\title{Find substring position (equivalent of PL/I index function)}
\usage{
cu_index(str, sub)
}
\arguments{
\item{str}{Character string to search in.}

\item{sub}{Character substring to search for.}
}
\value{
Integer position of first occurrence, or 0 if not found.
}
\description{
Find substring position (equivalent of PL/I index function)
}
\keyword{internal}
