% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_letbar-tree.R, R/cu_letbar.R
\name{cu_letbar}
\alias{cu_letbar}
\title{internal cufunction to construct significance letters on bars}
\usage{
cu_letbar(pvalue, psigcldnif, mapord, nlev, letleft = F, nletbarmax = 3)

cu_letbar(pvalue, psigcldnif, mapord, nlev, letleft = F, nletbarmax = 3)
}
\description{
internal cufunction to construct significance letters on bars

internal cufunction to construct significance letters on bars
}
\keyword{internal}
