% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_letbugfix.R
\name{cu_letbugfix}
\alias{cu_letbugfix}
\title{Fix reversed significance letters within factor levels}
\usage{
cu_letbugfix(letbar, namegroup1, gp1, gp2)
}
\arguments{
\item{letbar}{Character vector of significance letters.}

\item{namegroup1}{Name of the first grouping factor.}

\item{gp1}{Number of levels in factor 1.}

\item{gp2}{Number of levels in factor 2.}
}
\value{
Corrected character vector of significance letters.
}
\description{
Fix reversed significance letters within factor levels
}
\keyword{internal}
