% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_mapor2.R
\name{cu_mapor2}
\alias{cu_mapor2}
\title{internal cufunction used if bars are reordered with "order="}
\usage{
cu_mapor2(order, levnams, gp1, gp2, nlev)
}
\description{
internal cufunction used if bars are reordered with "order="
}
\keyword{internal}
