% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_mapord.R
\name{cu_mapord}
\alias{cu_mapord}
\title{internal cufunction used if bars are reordered with "order=" (defunct)}
\usage{
cu_mapord(order, levnams, nlev)
}
\description{
internal cufunction used if bars are reordered with "order=" (defunct)
}
\keyword{internal}
