% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_mapvalet.R
\name{cu_mapvalet}
\alias{cu_mapvalet}
\title{Map p-value to label or asterisks}
\usage{
cu_mapvalet(pval, pvlab, pvcuts)
}
\arguments{
\item{pval}{Numeric p-value.}

\item{pvlab}{Character, "p" for numeric or "*" for asterisks.}

\item{pvcuts}{Numeric vector of significance cut points.}
}
\value{
Formatted character string for the p-value.
}
\description{
Map p-value to label or asterisks
}
\keyword{internal}
