% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_mapvalst.R
\name{cu_mapvalst}
\alias{cu_mapvalst}
\title{Map p-value to asterisk notation}
\usage{
cu_mapvalst(pval, pvcuts)
}
\arguments{
\item{pval}{Numeric p-value.}

\item{pvcuts}{Numeric vector of significance cut points.}
}
\value{
Character string: "ns", "\emph{", "\strong{", "}}", or "****".
}
\description{
Map p-value to asterisk notation
}
\keyword{internal}
