% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_parsexp.R
\name{cu_parsexp}
\alias{cu_parsexp}
\title{internal function to get list of the names in chstr separated by + *}
\usage{
cu_parsexp(chstr)
}
\description{
internal function to get list of the names in chstr separated by + *
}
\keyword{internal}
