% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_plot2.R
\name{cu_plot2}
\alias{cu_plot2}
\title{internal cufunction to plot data by two factors}
\usage{
cu_plot2(
  progname,
  depvar,
  group1,
  group2,
  nlev1,
  nlev2,
  g1order,
  g1name,
  depname,
  g2name,
  title,
  plot = "bar",
  linetype = "n",
  linecolor = "black",
  linesize = 1,
  letbar = NULL,
  ebars = 4,
  dots = 1,
  barcolor = "black",
  barfill = "colors",
  pvdf = NULL,
  pvypos = NULL,
  pvstinc = NULL,
  pvlab = "p",
  pvprefix = "p=",
  pvsize = NULL,
  pvnshide = T,
  pvtipl = NULL,
  caption = NULL,
  legend = "top",
  size = NULL,
  width = NULL,
  posd = NULL,
  xangle = NULL,
  yangle = NULL,
  orientation = "vertical",
  binwfac = 30,
  dotsize = NULL,
  dotshape = NULL,
  dotcolor = NULL,
  vscale = "area",
  theme = "bw",
  yscale = "none",
  fontfamily = "sans",
  yrange = NULL,
  ymin = NA,
  ymax = NA,
  fontmain = NULL,
  fontxname = NULL,
  fontyname = NULL,
  fontxticks = NULL,
  fontyticks = NULL,
  axiscolor = NULL,
  tickcolor = NULL,
  axisthick = NULL,
  tickthick = NULL,
  ticklength = NULL,
  xticks.by = NULL,
  yticks.by = NULL,
  titlejust = "center",
  legheadsize = 12,
  legtextsize = 10,
  ftype = NULL,
  fname = NULL,
  fscale = NULL,
  fwidth = NULL,
  fheight = NULL,
  dpi = 300,
  remove = NULL,
  suff = NULL
)
}
\description{
internal cufunction to plot data by two factors
}
\keyword{internal}
