% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_plout.R
\name{cu_plout}
\alias{cu_plout}
\title{internal function to display plot object, write hires .ftype or .emf file}
\usage{
cu_plout(plotobj, funcname, ftype = NULL, fname = NULL, suff = NULL,
  seq = 0, scale = NULL, width = NULL, height = NULL, units = "in",
  dpi = 300, avplots = F, remove = NULL)
}
\description{
internal function to display plot object, write hires .ftype or .emf file
}
\keyword{internal}
