% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_pval9.R
\name{cu_pval9}
\alias{cu_pval9}
\title{internal cufunction to write p-value to console 9 chars}
\usage{
cu_pval9(pval)
}
\description{
internal cufunction to write p-value to console 9 chars
}
\keyword{internal}
