% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_pvalmatout.R
\name{cu_pvalmatout}
\alias{cu_pvalmatout}
\title{internal cufunction to print p.value with labels (in cucov2way)}
\usage{
cu_pvalmatout(p.value, levnams)
}
\description{
internal cufunction to print p.value with labels (in cucov2way)
}
\keyword{internal}
