% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_pvline.R
\name{cu_pvline}
\alias{cu_pvline}
\title{internal cufunction to prepare for p-value horizontals}
\usage{
cu_pvline(df, pvpairs, pvalues, nlev1, nlev2, pvlab, pnosig, psignif, p2stars,
  p3stars, p4stars, letbar, pvypos, pvnshide, pvspill, chpvref, ebars, dots,
  yscale)
}
\description{
internal cufunction to prepare for p-value horizontals
}
\keyword{internal}
