% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_reorder.R
\name{cu_reorder}
\alias{cu_reorder}
\title{Reorder factor levels}
\usage{
cu_reorder(group, order)
}
\arguments{
\item{group}{Factor variable to reorder.}

\item{order}{Character vector specifying new level order, or NULL.}
}
\value{
Factor with reordered levels.
}
\description{
Reorder factor levels
}
\keyword{internal}
