% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_rep2fac.R
\name{cu_rep2fac}
\alias{cu_rep2fac}
\title{Does repeated measures anova and post-hoc tests when there is a second, non-repeated factor (long-format - multiple lines for each subject)}
\usage{
cu_rep2fac(depvar, group1, group2, Subject, interact = TRUE,
  partialF = FALSE, mainx = 1, scale = "frequency", ebars = 1,
  dots = 0, barcolor = "black", barfill = "colors", plot = "bar",
  order = NULL, psigcld = 0, conf.int = 0.95, depname = NULL,
  g1name = NULL, g2name = NULL, pvpairs = "all", pvypos = NULL,
  pvsi = 0.05, pvlab = "p", pvprefix = "p=", chpvref = "ref",
  pvcuts = c(0.2, 0.05, 0.01, 0.001), pvnshide = T, pvtipl = 0.01,
  title = NULL, legend = "top", suff = NULL, caption = NULL,
  linetype = "n", linecolor = "black", linesize = 1, theme = "bw",
  size = NULL, width = NULL, yscale = "none", font.x = NULL,
  xangle = NULL, yangle = NULL, orientation = "vertical", posd = NULL,
  binwfac = 30, dotsize = NULL, dotshape = NULL, dotcolor = NULL,
  ftype = NULL, fname = NULL, fscale = NULL, fwidth = NULL,
  fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{depvar}{, group1, group2, Subject  (4 required)}

\item{group1}{repeated grouping factor}

\item{interact}{=TRUE (default), F or FALSE for no interaction}

\item{partialF}{=FALSE (default), T or TRUE for partial F vs simpler models}

\item{mainx}{=1 (default), 2 to interchange the two factors in bar graph}

\item{scale}{="frequency" (default) or "percent"}

\item{ebars}{=1 (default)/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR)}

\item{dots}{=0 (default), 1 to display data on graph}

\item{barfill}{="colors" (default) for colors by group ("lancet" for 2 factors, other journal options "aaas", "jco", "uchicago", "npg"}

\item{plot}{="bar" (default) for bar graphs; "box" "violin" "rod" "no"}

\item{order}{=NULL (default)/c("...") to reorder bars}

\item{psigcld}{=0 (no letters)/x for CLD letters on bars (any shared letter means P>psigcld, 0 for no letters)}

\item{conf.int}{=0.95 (default)/x/0 for confidence interval width of contrast estimates (0 for none)}

\item{depname}{/g1name/g2name/title to override names of depvar, groupvar, title}

\item{legend}{="top" (default), "bottom", "right", "left" to locate legend}

\item{linetype}{="n" (default)/x for no connecting lines ("dashed" "dotted")}

\item{linesize}{=1 (default)/x for line thickness}

\item{theme}{="bw" (default)/x for white background (or "gray" or "classic" (no grid lines))}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can change to "horizontal"}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=NULL (default 30) set to fraction of range within which points will be binned}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{dotcolor}{=NULL (default "white") set to dot color}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Does repeated measures anova and post-hoc tests when there is a second, non-repeated factor (long-format - multiple lines for each subject)
}
\examples{
# Internal function, not exported; called via curepmeas()
\dontrun{
cu_rep2way(TG, Diet, sex, ID)  # with interaction between the two fixed factors
# with no interaction (pointless, same result as without 2nd factor)
cu_rep2way(TG, Diet, sex, ID, interact=FALSE)
}
}
