% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_sprintstr.R
\name{cu_sprintstr}
\alias{cu_sprintstr}
\title{internal cufunction to create a string of length width, padded or truncated}
\usage{
cu_sprintstr(chstr, width)
}
\description{
internal cufunction to create a string of length width, padded or truncated
}
\keyword{internal}
