% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_table0.R
\name{cu_table0}
\alias{cu_table0}
\title{internal cufunction to calculate summary stats for variables}
\usage{
cu_table0(
  depvar,
  group1 = NULL,
  brief = FALSE,
  sdsamerow = FALSE,
  doAll = FALSE,
  pnorm = 0.05,
  chariqr = ":",
  nlevmax = 9
)
}
\description{
internal cufunction to calculate summary stats for variables
}
\keyword{internal}
