% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_test.R
\name{cu_test}
\alias{cu_test}
\title{Test function for internal development}
\usage{
cu_test(val)
}
\arguments{
\item{val}{Test input value.}
}
\value{
A list with test objects.
}
\description{
Test function for internal development
}
\keyword{internal}
