% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cucov1way.R
\name{cucov1way}
\alias{cucov1way}
\title{Does one-way analysis of covariance (ancova), all contrasts, and a graph}
\usage{
cucov1way(depvar, covar, group1, xs = NULL, breakpt = F, ebars = 1,
  dosimpler = T, partialF = TRUE, plotcons = "yes", depname = NULL,
  covname = NULL, g1name = NULL, title = NULL, caption = "",
  g1order = NULL, conf.int = 0.95, color = "color", padj = "none",
  pbart = 0.05, pnorm = 0.05, chariqr = ",", legend = "top",
  xangle = NULL, yangle = NULL, orientation = "vertical",
  binwidth = NULL, shape = 16, dotcolor = "black", dotsize = 2,
  linetype = "solid", linesize = 1, linecolor = "red", theme = "bw",
  fontfamily = "sans", xmin = NA, ymin = NA, xmax = NA, ymax = NA,
  fontmain = NULL, fontxname = NULL, fontyname = NULL,
  fontxticks = NULL, fontyticks = NULL, axiscolor = NULL,
  tickcolor = NULL, axisthick = NULL, tickthick = NULL,
  ticklength = NULL, xticks.by = NULL, yticks.by = NULL,
  titlejust = "center", legheadsize = 12, legtextsize = 10,
  ftype = NULL, fname = NULL, fscale = NULL, fwidth = NULL,
  fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{depvar}{,covar,group1 required: dep-var, covariate, grouping factor}

\item{xs}{optional, c(x1,x2,..) to model with interaction and do contrasts at x1,x2,...}

\item{breakpt}{=F (default) or value at which covar slope changes}

\item{ebars}{=1 (default)/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR)}

\item{dosimpler}{=T (default) or F to do or not do cu1way first}

\item{partialF}{=TRUE (default), F or FALSE for no partial F vs simpler models}

\item{plotcons}{="yes" (default), "plot" for just plot, "cons" for just console}

\item{depname}{/covname/g1name/title to override names of depvar, covar, group1, title}

\item{caption}{='' (default) or set to string to show at bottom right}

\item{g1order}{=NULL (default)/c("...") to reorder group1}

\item{conf.int}{=0.95 (default) for confidence interval width of contrast estimates (0 for none)}

\item{color}{="color" (default) for color graph, "black" for black}

\item{padj}{="none"(default), p-value adjustment for multiple comparisons, can be "bonferroni", "holm")}

\item{pbart}{=0.05 (default)/x for Bartlett test threshold for variance homogeneity}

\item{pnorm}{=0.05 (default)/x for normality test threshold}

\item{chariqr}{="," (default)/x for character to separate quartiles ("-" etc)}

\item{legend}{(default="top"), can be "bottom", "right", "left"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can change to "horizontal"}

\item{shape}{=16 (default closed circle) (see http://www.sthda.com/english/wiki/ggplot2-point-shapes)}

\item{dotcolor}{="black" (default) for point symbol color (="red" ="blue" etc)}

\item{dotsize}{=2 (default) or x to set size of point symbols}

\item{linetype}{="solid"(default)/x for solid line ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linesize}{=1 (default)/x for line thickness}

\item{linecolor}{="red" (default)/x for line color}

\item{theme}{="bw" (default)/x for white background ("classic" (no grid lines),"linedraw" "gray" "minimal" "void")}

\item{fontfamily}{="sans" (default), can be "serif" "mono"}

\item{xmin}{,xmax,ymin,ymax =NA (default) or value to start/end x/y-axis}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for not title}

\item{fontxname}{.fontyname,fontxticks,fontyticks = c(12,"plain","black") default, 0 to suppress}

\item{axiscolor}{,tickcolor="black" (default)/x for axis/tick color}

\item{axisthick}{,tickthick=0.5 (default)/x for axis/tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{,yticks.by =NULL (default)/s for x/y tick spacing by s}

\item{titlejust}{="center" (default) or "left" or "right"}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Does one-way analysis of covariance (ancova), all contrasts, and a graph
}
\examples{
# for regression, plot, contrasts (parallel lines, no interaction)
with(NEJM, cucov1way(tcstudy, tcpre, Diet))
# for regression, plot, contrasts at tcpre=150,220 (non-parallel lines)
with(NEJM, cucov1way(tcstudy, tcpre, Diet, c(150, 220)))
}
