% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cucov2way.R
\name{cucov2way}
\alias{cucov2way}
\title{Does two-way analysis of covariance (ancova), all contrasts, and a graph}
\usage{
cucov2way(depvar, covar, group1, group2, xs = NULL, interact = T,
  dosimpler = T, partialF = TRUE, depname = NULL, covname = NULL,
  g1name = NULL, g2name = NULL, title = NULL, caption = "",
  g1order = NULL, g2order = NULL, conf.int = 0.95, ebars = 1,
  dots = 0, color = "color", padj = "none", pbart = 0.05,
  pnorm = 0.05, chariqr = ",", legend = "top", xangle = NULL,
  yangle = NULL, orientation = "vertical", posd = NULL, shape = 16,
  dotcolor = "black", dotsize = 2, linetype = "solid", linesize = 1,
  linecolor = "red", theme = "bw", fontfamily = "sans", xmin = NA,
  ymin = NA, xmax = NA, ymax = NA, fontmain = NULL, fontxname = NULL,
  fontyname = NULL, fontxticks = NULL, fontyticks = NULL,
  axiscolor = NULL, tickcolor = NULL, axisthick = NULL,
  tickthick = NULL, ticklength = NULL, xticks.by = NULL,
  yticks.by = NULL, titlejust = "center", legheadsize = 12,
  legtextsize = 10, ftype = NULL, fname = NULL, fscale = NULL,
  fwidth = NULL, fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{depvar}{,covar, group1, group2 required: dep-var, covariate, grouping factors 1 & 2}

\item{xs}{optional, c(x1,x2,..) to model with interaction and do contrasts at x1,x2,...}

\item{interact}{=TRUE (default), F or FALSE for no group1*group2 interaction}

\item{dosimpler}{=T (default) or F to do or not do cu2way first}

\item{partialF}{=TRUE (default), F or FALSE for no partial F vs simpler models}

\item{depname}{/covname/g1name/g2name/title to override names of depvar, covar, group1, group2, title}

\item{caption}{='' (default) or set to string to show at bottom right}

\item{g1order}{=NULL (default) to reorder group1 (1st Factor)}

\item{g2order}{=NULL (default) to reorder group2 (2nd Factor)}

\item{conf.int}{=0.95 (default) for confidence interval width of contrast estimates (0 for none)}

\item{ebars}{=1 (default)/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR)}

\item{dots}{=0 (default), 1 to display data on graph}

\item{color}{="color" (default) for color graph, "black" for black}

\item{padj}{="none"(default), p-value adjustment for multiple comparisons, can be "bonferroni", "holm")}

\item{pbart}{=0.05 (default)/x for Bartlett test threshold for variance homogeneity}

\item{pnorm}{=0.05 (default)/x for normality test threshold}

\item{chariqr}{="," (default)/x for character to separate quartiles ("-" etc)}

\item{legend}{(default="top"), can be "bottom", "right", "left"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can change to "horizontal"}

\item{posd}{NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{shape}{=16 (default closed circle) (see http://www.sthda.com/english/wiki/ggplot2-point-shapes)}

\item{dotcolor}{="black" (default) for point symbol color (="red" ="blue" etc)}

\item{dotsize}{=2 (default) or x to set size of point symbols}

\item{linetype}{="solid"(default)/x for solid line ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linesize}{=1 (default)/x for line thickness}

\item{linecolor}{="red" (default)/x for line color}

\item{theme}{="bw" (default)/x for white background ("classic" (no grid lines),"linedraw" "gray" "minimal" "void")}

\item{fontfamily}{="sans" (default), can be "serif" "mono"}

\item{xmin}{,xmax,ymin,ymax =NA (default) or value to start/end x/y-axis}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for not title}

\item{fontxname}{.fontyname,fontxticks,fontyticks = c(12,"plain","black") default, 0 to suppress}

\item{axiscolor}{,tickcolor="black" (default)/x for axis/tick color}

\item{axisthick}{,tickthick=0.5 (default)/x for axis/tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{,yticks.by =NULL (default)/s for x/y tick spacing by s}

\item{titlejust}{="center" (default) or "left" or "right"}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Does two-way analysis of covariance (ancova), all contrasts, and a graph
}
\examples{
attach(NEJM)
cucov2way(tcstudy, tcpre, Diet, sex)  # regression, plot, contrasts (parallel lines, Diet*sex interaction)
cucov2way(tcstudy, tcpre, Diet, sex, interact=FALSE)  # regression, plot, contrasts (parallel lines, no interaction)
# regression, plot, contrasts at tcpre=150,220 (non-parallel lines, tcpre*Diet*sex interaction)
cucov2way(tcstudy, tcpre, Diet, sex, c(150, 220))
# regression, plot, contrasts at tcpre=150,220 (non-parallel lines, no interaction)
cucov2way(tcstudy, tcpre, Diet, sex, interact=FALSE, c(150, 220))
detach(NEJM)
}
