% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cucox.R
\name{cucox}
\alias{cucox}
\title{Cox modeling, all possible models, summarize best (or specified) model,
LR tests, ROCs, K-M of Cox score strata, predicted K-M for each factor}
\usage{
cucox(dsgiven, timnam, depnam, formula, dopredkm = T, doroc = T,
  docoxkm = T, wtnam = NULL, printfit = F, dodredge = T, usemod = 1,
  nmodshow = 16, m.min = 1, m.max = 99, fixed = NULL,
  subset = as.expression(1), tmax = 0, xlabpredkm = "Time",
  ylabpredkm = "Survival", titlepredkm = "Predicted Kaplan-Meier",
  palette = "hue", fun = "pct", color = "red",
  xlabroc = "1 - Specificity", ylabroc = "Sensitivity", allkmpairs = F,
  xlabkm = NULL, ylabkm = NULL, titlekm = NULL, nverkm = 2,
  nhorkm = 2, nverroc = 2, nhorroc = 2, ftype = NULL, fname = NULL,
  fscale = NULL, fwidth = NULL, fheight = NULL, dpi = 300,
  remove = NULL)
}
\arguments{
\item{dsgiven}{, timnam, depnam, formula  required: dataset, char-strings of names of time-var and variable being analyzed, and model formula}

\item{dopredkm}{=T (default) to get (T) or not get (F) predicted Kaplan-Meier curves}

\item{doroc}{=T (default, at 50,75,100\% events) for ROC curves [c(,,) list to specify times]}

\item{docoxkm}{=T (default, quartiles) for K-M vs Cox risk score [c(,,) list to specify cut-pts]}

\item{wtnam}{=NULL (default) char-string pointing to variable with weights}

\item{printfit}{=F (default) to get (T) or not get (F) fitted risk scores for all subjects}

\item{dodredge}{=T (default) to get (T) or not get (F) all possible models}

\item{usemod}{=1 (default) to use the usemod-th best found by dredge}

\item{nmodshow}{=16 (default) to show top nmodshow models by dredge}

\item{m.min}{and m.max= to set smallest and largest model sizes in dredge}

\item{fixed}{= c(,,) list to specify variables that must be in model}

\item{subset}{=logical expression describing models to keep in dredge}

\item{tmax}{=0 (default for no max T), numeric study end time}

\item{xlabpredkm}{="Time" (default) for predicted K-M x-axis label}

\item{ylabpredkm}{="Survival" (default) for predicted K-M y-axis label}

\item{titlepredkm}{="Predicted Kaplan-Meier Curves" (default) for predicted K-M title}

\item{palette}{="hue" (default) "grey" "RdBu" "Blues" "lancet" etc}

\item{fun}{="pct" (default) for survival, "event" for cumulative events, "cumhaz" for cumulative hazard}

\item{color}{="red" (default) for single ROC curve color (glm object)}

\item{xlabroc}{="x axis label" (default "1 - Specificity")}

\item{ylabroc}{="y axis label" (default "Sensitivity")}

\item{allkmpairs}{=T (default) or F (to not compare each pair of curves)}

\item{xlabkm}{=NULL (default) for K-M x-axis label}

\item{ylabkm}{=NULL (default) for K-M y-axis label}

\item{titlekm}{=NULL (default) for K-M title}

\item{nverkm, nhorkm, nverroc, nhorroc}{=2 (default) #km,rocplots vert,horiz}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Cox modeling, all possible models, summarize best (or specified) model,
LR tests, ROCs, K-M of Cox score strata, predicted K-M for each factor
}
\examples{
cucox(coxdata, "TimeToEvent", "Outcome", "C.Index + scale(LVEF) + scale(BNP)")
}
