% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cukm.R
\name{cukm}
\alias{cukm}
\title{Kaplan-Meier curves with p-value, # at risk}
\usage{
cukm(timevar, statvar, ttmtvar, kmtype = "survival", pvalue = T,
  allkmpairs = T, tmax = 0, xlab = "Time", ylab = NULL, title = NULL,
  rho = 0, legend = "top", legend.labs = NULL, risk.table.y.text = F,
  scale = "default", linetype = 1, censor = TRUE, censor.shape = "+",
  ftype = NULL, fname = NULL, fscale = NULL, fwidth = NULL,
  fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{timevar}{, statvar, ttmtvar (3 required)}

\item{kmtype}{="survival" (default) or "ci" (cumulative incidence)}

\item{pvalue}{=T (default) or F (for no p-value)}

\item{allkmpairs}{=T (default) or F (to not compare each pair of curves)}

\item{tmax}{=0 (default for no max T), numeric study end time}

\item{xlab}{/ylab/title=NULL to override labels of x-axis/y-axis/title}

\item{rho}{=0 (log-rank default) or 1 (Peto-Peto)}

\item{legend}{="top" (default), "bottom", "right", "left" to locate legend}

\item{legend.labs}{=NULL(default)/c(...) to override risk factor level names}

\item{risk.table.y.text}{=F (default) =T to list level names in risk table}

\item{scale}{="percent" to have y-axis display percent instead of decimal (default)}

\item{linetype}{=1 (default for solid), 2 (dashed), 3 (dotted)}

\item{censor}{=T (default) or F to show (or not) censoring on curves}

\item{censor.shape}{="+" (default) or any character to indicate censoring}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Does K-M curves of 2nd argument vs 1st for different levels of 3rd argument.
Displays table of # at risk, and p-value. Does survival (default) or cumulative incidence.
Plus sign indicates loss to follow-up (right-censoring).
}
\examples{
attach(hepc)
cukm(time, status, Treatment)
cukm(time, status, Treatment, kmtype="ci")  # for cumulative incidence
detach(hepc)
}
