% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culinreg.R
\name{culinreg}
\alias{culinreg}
\title{Multiple linear regression, all possible models, summarize best (or specified) model,
do added variable plots}
\usage{
culinreg(
  dsgiven,
  depnam,
  formula,
  weights = NULL,
  printfit = F,
  ndecfit = 2,
  dodredge = T,
  usemod = 1,
  nmodshow = 16,
  m.min = 1,
  m.max = 99,
  fixed = NULL,
  subset = as.expression(1),
  fitasx = T,
  ftype = NULL,
  fname = NULL,
  fscale = NULL,
  fwidth = NULL,
  fheight = NULL,
  dpi = 300,
  remove = NULL
)
}
\arguments{
\item{dsgiven}{, depnam, formula required: dataset, char-string of name of variable being analyzed,}

\item{formula}{of model "preda+predb+predc*predd" etc. Add "+0" for zero intercept}

\item{printfit}{=F (default) to get (T) or not get (F) fitted values for all subjects}

\item{ndecfit}{=3 (default) for the number of decimal digits in fitted values}

\item{dodredge}{=T (default) to get (T) or not get (F) all possible models}

\item{usemod}{=1 (default) to use the usemod-th best found by dredge}

\item{nmodshow}{=16 (default) to show top nmodshow models by dredge}

\item{m.min}{and m.max= to set smallest and largest model sizes in dredge}

\item{fixed}{= c(,,,) to specify variables that must be in model}

\item{subset}{=logical expression describing models to keep in dredge}

\item{fitasx}{=T(default) or set to T to plot observed as x}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Multiple linear regression, all possible models, summarize best (or specified) model,
do added variable plots
}
\examples{
\dontrun{
culinreg(NEJM, "tcstudy", "tcpre+tgpre+Diet")
}
}
