% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culogist.R
\name{culogist}
\alias{culogist}
\title{Logistic regression, all possible models, summarize best (or specified) model,
do ROC curve(s), AUC, and summary table with best accuracy}
\usage{
culogist(dsgiven, depnam, formula, xs = NULL, ordinal = NULL,
  logitlog = "logit", start = NULL, wtnam = NULL, dosimpler = T,
  nlevmax = 5, printfit = F, ndecfit = 1, dodredge = T, usemod = 1,
  nmodshow = 16, m.min = 1, m.max = 99, fixed = NULL,
  subset = as.expression(1), emf = F, xlab = "1 - Specificity",
  ylab = "Sensitivity", color = "red", veccolor = c("red", "blue",
  "green", "black"))
}
\arguments{
\item{dsgiven}{, depnam, formula required: dataset, char-strings of name of variable being analyzed, and model formula}

\item{xs}{list of probability cut points for 2x2 tables (besides 0.5)}

\item{ordinal}{=NULL (default)/c("...") to specify ordering of variable being analyzed}

\item{logitlog}{="logit" (default) to model odds ("logit") or risk ("log")}

\item{start}{=NULL (default) a list of initial guesses at model coefficients}

\item{wtnam}{=NULL (default) char-string pointing to variable with weights}

\item{dosimpler}{=T (default) to compare (T) or not (F) each indep-var across levels of dep-var}

\item{nlevmax}{=5 (default) for max # of levels in dependent variable}

\item{printfit}{=F (default) to get (T) or not get (F) fitted probabilities for all subjects}

\item{ndecfit}{=3 (default) for the number of decimal digits in fitted probabilities}

\item{dodredge}{=T (default) to get (T) or not get (F) all possible models}

\item{usemod}{=1 (default) to use the usemod-th best found by dredge}

\item{nmodshow}{=16 (default) to show top nmodshow models by dredge}

\item{m.min}{and m.max= to set smallest and largest model sizes in dredge}

\item{fixed}{=c(,,,) to specify variables that must be in model}

\item{subset}{logical expression describing models to keep in dredge}

\item{emf}{=F(default)/T/name (to create culogist.emf or name.emf needed on Mac)}

\item{xlab}{="x axis label" (default "1 - Specificity")}

\item{ylab}{="y axis label" (default "Sensitivity")}

\item{color}{="red" (default) for ROC curve color}

\item{veccolor}{=c("red","blue","green","black") (default) for multiple curves (polr object)}
}
\value{
returns nothing
}
\description{
Logistic regression, all possible models, summarize best (or specified) model,
do ROC curve(s), AUC, and summary table with best accuracy
}
\examples{
# for AUC and summary table with best accuracy
culogist(AJCN, "MetSyn", "TG+HDL+LN_TG+BMI+GLUC")
# 3 summary tables (best accuracy, prob=0.3, 0.7)
culogist(AJCN, "MetSyn", "TG+HDL+LN_TG+BMI+GLUC", c(0.3, 0.7))
}
