% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuomics.R
\name{cuomics}
\alias{cuomics}
\title{Analyzes multiple outcomes by specified function name}
\usage{
cuomics(funcname, dsgiven, dnamf, dnaml, fnom, ...)
}
\arguments{
\item{dsgiven}{,dnamf,dnaml required: dataset, char-strings of names of first&last variable being analyzed}

\item{fnom}{= string as prefix for output file cuomics.csv}

\item{...}{remaining arguments for chosen function}

\item{funcname:}{"1way","2way","cov1way","cov2way","scatter","repmeasl","repmeasw"}
}
\value{
nothing
}
\description{
Analyzes multiple outcomes by specified function name
}
\examples{
\dontrun{
cuomics("1way", NEJM, "tcstudy", "hcchange", "tchc1w", Diet)
cuomics("2way", NEJM, "tcstudy", "hcchange", "tchc2w", Diet, sex)
cuomics("cov1way", NEJM, "tcstudy", "hcchange", "tchccov1", tcpre, Diet)
cuomics("cov2way", NEJM, "tcstudy", "hcchange", "tchccov2", tcpre, Diet, sex)
cuomics("scatter", NEJM, "tcstudy", "hcchange", "tchcscat", tcpre)
cuomics("repmeasw", delta, "TC", "like", "tclikerep", "Diet", "sex")
}
}
