% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupairwise.t.R
\name{cupairwise.t}
\alias{cupairwise.t}
\title{pairwise.t.test to handle group(s) with just 1 observation
same arguments as pairwise.t.test, same code (Nov.20) with added:
for (j in 1:length(s)) {if (is.na(s[j])) s[j] <- 0}}
\usage{
cupairwise.t(x, g, p.adjust.method = p.adjust.methods, pool.sd = !paired,
  paired = FALSE, alternative = c("two.sided", "less", "greater"), ...)
}
\description{
pairwise.t.test to handle group(s) with just 1 observation
same arguments as pairwise.t.test, same code (Nov.20) with added:
for (j in 1:length(s)) {if (is.na(s[j])) s[j] <- 0}
}
