% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupartialF.R
\name{cupartialF}
\alias{cupartialF}
\title{partial F test to test if model augmentation is significant}
\usage{
cupartialF(rse1, ndf1, rse2, ndf2)
}
\arguments{
\item{rse1}{residual standard error of reduced (simpler) model (required)}

\item{ndf1}{degrees of freedom of reduced (simpler) model (required)}

\item{rse2}{residual standard error of full (bigger) model (required)}

\item{ndf2}{degrees of freedom of full (bigger) model (required)}
}
\value{
returns p-value of partial F(np,ndf2) = ((rse1^2*ndf1-rse2^2*ndf2)/np)/rse2^2
}
\description{
partial F test to test if model augmentation is significant
}
\examples{
cupartialF(12.39, 33, 7.56, 30)
}
