% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curead.R
\name{curead}
\alias{curead}
\title{Does a read.table and returns a dataframe object}
\usage{
curead(skip = NULL, nrows = -1, maxfaclev = 0, maxvarprint = 20)
}
\arguments{
\item{skip}{=lines to be skipped (default=NULL)}

\item{nrows}{=lines to be read after line with names (default is all)}

\item{maxfaclev}{if < levels, treated as char-string (default sqrt(n))}

\item{maxvarprint}{=20 (default), or N, with no print if #variables larger}
}
\value{
dataframe object read
}
\description{
Does a read.table and returns a dataframe object
}
\examples{
# Requires interactive file chooser dialog
\dontrun{
HD = curead()
NEJM = curead(skip=1)
}
}
