% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curepmeas.R
\name{curepmeas}
\alias{curepmeas}
\title{Does repeated measures one- or two-way anova and post-hoc tests (format wide or long - 1/many lines for each subject)}
\usage{
curepmeas(
  dsgiven,
  dnam,
  repnam,
  fac2 = NULL,
  idnam = NULL,
  minimal = F,
  interact = TRUE,
  partialF = TRUE,
  cov = NULL,
  ytrans = "none",
  maxlines = 0,
  mainx = 1,
  sumdiff = 1,
  scale = "frequency",
  scatter = "yes",
  ebars = 1,
  dots = 0,
  barcolor = "black",
  barfill = NULL,
  plot = "bar",
  ordinal = NULL,
  g1order = NULL,
  g2order = NULL,
  difforder = NULL,
  depname = NULL,
  g1name = NULL,
  g2name = NULL,
  title = NULL,
  pnorm = 0.05,
  psigcld = 0,
  conf.int = 0.95,
  ddepn = NULL,
  shape = 16,
  caption = NULL,
  pvpairs = "std",
  pvypos = NULL,
  pvstinc = 0.05,
  pvlab = "p",
  pvprefix = "p=",
  pvsize = NULL,
  chpvref = "ref",
  pvspill = F,
  pnosig = 0.2,
  psignif = 0.05,
  p2stars = 0.01,
  p3stars = 0.001,
  p4stars = 1e-04,
  pvnshide = T,
  pvtipl = 0.01,
  ifdeb = F,
  refmean = NA,
  normvisit = 0,
  normway = "fold",
  rmwall = NULL,
  needvar = 0,
  linetype = "n",
  linecolor = "black",
  linesize = 1,
  theme = "bw",
  legend = "top",
  size = NULL,
  width = NULL,
  yscale = "none",
  xangle = NULL,
  yangle = NULL,
  orientation = "vertical",
  posd = NULL,
  binwfac = 30,
  dotsize = NULL,
  dotshape = NULL,
  dotcolor = NULL,
  fontmain = NULL,
  fontxname = NULL,
  fontyname = NULL,
  fontxticks = NULL,
  fontyticks = NULL,
  axiscolor = NULL,
  tickcolor = NULL,
  axisthick = NULL,
  tickthick = NULL,
  ticklength = NULL,
  xticks.by = NULL,
  yticks.by = NULL,
  dodredge = T,
  usemod = 1,
  nmodshow = 16,
  nverscat = 2,
  nhorscat = 2,
  m.min = 1,
  m.max = 99,
  fixed = NULL,
  subset = as.expression(1),
  ftype = NULL,
  fname = NULL,
  fscale = NULL,
  fwidth = NULL,
  fheight = NULL,
  dpi = 300,
  remove = NULL
)
}
\arguments{
\item{dsgiven}{,dnam,repnam required: dataset, char-strings of names of variable being analyzed, repeated grouping factor}

\item{fac2}{char-string of name of 2nd, non-repeated grouping factor (default NULL)}

\item{idnam}{(to indicate long format) =char-string of name of ID/Subject variable (default NULL)}

\item{minimal}{= FALSE (default), T or TRUE to suppress all output (for cuomics)}

\item{interact}{=TRUE (default), F or FALSE for no interaction}

\item{partialF}{=TRUE (default), F or FALSE for no partial F vs simpler models}

\item{cov}{=char-string of covariate model (default NULL)}

\item{ytrans}{="none" (default) to transform depvar: "sqrt" "log"}

\item{maxlines}{=0 (default) for no line plot, 1 for line plots, n to limit #lines on each to n}

\item{mainx}{=1 (default), 2 to interchange the two factors in bar graph}

\item{sumdiff}{=1 (default) or -1 for difference summary and bar graph, 0 for not (1 for rest-Igroup, -1 for Igroup-rest)}

\item{scale}{="frequency" (default) or "percent"}

\item{scatter}{= "yes" (default) for scatter plots of repeated measures, "no" for none}

\item{ebars}{=1 (default)/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR)}

\item{dots}{=0 (default), 1 to display data on graph}

\item{barcolor}{for outline color, barfill for fill color. Use barfill="colors" for colors by group.}

\item{plot}{="bar" (default) for bar graphs; "box" "violin" "rod" "no"}

\item{ordinal}{=NULL (default)/c("...") to reorder dependent variable}

\item{g1order}{,g2order, difforder=NULL (defaults)/c("...") to reorder bars (I Factor, Differences)}

\item{depname}{/g1name/g2name/title to override names of depvar, groupvar, fac2, title}

\item{pnorm}{=0.05 (default) threshold for normality test}

\item{psigcld}{=0 (no letters)/x for CLD letters on bars (any shared letter means P>psigcld, 0 for no letters)}

\item{conf.int}{=0.95 (default)/x/0 for confidence interval width of contrast estimates (0 for none)}

\item{ddepn}{=NULL (default) for name of dependent variable difference}

\item{shape}{=16 (default)/n for shape of dots in regression plot}

\item{caption}{=NULL (default)/"yes"/"caption text" to get caption ("yes" to list n's)}

\item{pnosig}{=0.2 (default)/x for threshold to show p-values even if nonsigificant}

\item{psignif}{=0.05 (default)/x for threshold to significance}

\item{p2stars}{=0.01 (default)/x for threshold to two stars}

\item{p3stars}{=0.001 (default)/x for threshold to three stars}

\item{p4stars}{=0.0001 (default)/x for threshold to four stars}

\item{ifdeb}{=F (default)/T to suppress debug output or not}

\item{refmean}{=NA (default)/ref.val to compare each group mean with}

\item{normvisit}{=0 (default)/i to specify the i-th visit as the reference}

\item{normway}{="fold" (default), "sub","pct" to indicate how data get normalized to normvisit value}

\item{rmwall}{=NULL set to dataset to be updated for writing out by cuomics}

\item{needvar}{=0/the number of other variables to include in rmwall}

\item{linetype}{="n" (default)/x for no connecting lines ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linecolor}{="black" (default)/x for black lines ("red" etc) (only for single factor)}

\item{linesize}{=1 (default)/x for line thickness}

\item{theme}{="bw" (default)/x for white background (or "gray" or "classic" (no grid lines))}

\item{legend}{="top" (default), "bottom", "right", "left" to locate legend}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can be "horizontal" or "reverse"}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=NULL (default 30) set to fraction of range within which points will be binned}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{dotshape}{=NULL (default)}

\item{dotcolor}{=NULL (default "white") set to dot color}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for not title}

\item{fontxname}{.fontyname,fontxticks,fontyticks = c(12,"plain","black") default, 0 to suppress}

\item{axiscolor}{,tickcolor="black" (default)/x for axis/tick color}

\item{axisthick}{,tickthick=0.5 (default)/x for axis/tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{,yticks.by =NULL (default)/s for x/y tick spacing by s}

\item{dodredge}{=T (default) to get (T) or not get (F) all possible models}

\item{usemod}{=1 (default) to use the usemod-th best found by dredge}

\item{nmodshow}{=16 (default) to show top nmodshow models by dredge}

\item{nverscat, nhorscat}{=2 (default) #scatterplots vert,horiz}

\item{m.min}{and m.max= to set smallest and largest model sizes in dreedge}

\item{fixed}{to specify variables that must be in model}

\item{subset}{logical expression describing models to keep in dredge}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}

\item{pvpairs="std"}{(default)/"all"/i/c(ij,ik,...) to show std or all or vs.i or pval's of i/j, i/k ...}

\item{pvypos=NULL}{(default)/position of pval lines}

\item{pvstinc=0.05}{(default)/increment for pvypos from one pval line to next}

\item{pvlab="p"/"*"}{to display numerical or asterisks}

\item{pvprefix="p="}{(default)/"" etc prefix to p-values}

\item{pvsize=NULL}{(default 3.5, 7)/size of p-values or **}

\item{chpvref="ref"}{(default) char above i-th group bar when pvpairs=i}

\item{pvspill=F/T}{to not allow p-values to spill outside range or to allow}

\item{pvnshide=T}{(default)/F to hide NS p-values or not}

\item{pvtipl=0.01}{(default)/length of p-value line tips}
}
\value{
one line of summary means/SDs and p-values, data if normvisit
}
\description{
Does repeated measures one- or two-way anova and post-hoc tests (format wide or long - 1/many lines for each subject)
}
\examples{
\dontrun{
curepmeas(delta,"TG","Diet")  # one line per subject; TG.xx TG.yy etc as names of TG data
curepmeas(delta,"TG","Diet", "sex")  # with sex as second factor, Diet*sex interaction in model
curepmeas(delta,"TG","Diet", "sex", interact=F)  # no Diet*sex interaction
curepmeas(delta,"TG","Diet", "sex", cov="age+sex*age")  # with covariate adjustment
}
}
