% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.data.frame.R
\name{cusum.data.frame}
\alias{cusum.data.frame}
\title{Data frame method for cusum}
\usage{
\method{cusum}{data.frame}(object, maxsum = 7L, digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{Data frame to summarize.}

\item{maxsum}{Maximum number of levels to show for factors.}

\item{digits}{Number of significant digits.}

\item{...}{Additional arguments.}
}
\value{
Summary table.
}
\description{
Data frame method for cusum
}
