% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.default.R
\name{cusum.default}
\alias{cusum.default}
\title{Default method for cusum}
\usage{
\method{cusum}{default}(object, ..., digits = max(3L, getOption("digits") -
  3L))
}
\arguments{
\item{object}{Object to summarize.}

\item{...}{Additional arguments.}

\item{digits}{Number of significant digits.}
}
\value{
Summary statistics.
}
\description{
Default method for cusum
}
