% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\name{cusum}
\alias{cusum}
\title{Does summary stats}
\usage{
cusum(object, ...)
}
\arguments{
\item{object}{to be summarized}

\item{...}{additional arguments passed to methods}
}
\value{
returns nothing
}
\description{
Does summary stats
}
