% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutable1.R
\name{cutable1}
\alias{cutable1}
\title{Table 1 with optional box or violin plots}
\usage{
cutable1(ds, group1 = NULL, group2 = NULL, doAll = T, brief = F,
  pnorm = 0.05, sdsamerow = T, compare = F, minimal = F, docorr = F,
  plot = "box", ebars = NULL, dots = 1, width = NULL,
  barcolor = "black", barfill = "colors", depname = NULL,
  g1name = NULL, g2name = NULL, title = NULL, g1order = NULL,
  g2order = NULL, caption = NULL, xlab = NULL, ylab = NULL,
  chariqr = ":", charamp = "&", nlevmax = 9, pvpairs = "std",
  pvypos = NULL, pvsi = 0.05, pvlab = "p", pvprefix = "p=",
  pvsize = 3.5, chpvref = "ref", pvspill = F, pnosig = 0.2,
  psignif = 0.05, p2stars = 0.01, p3stars = 0.001, p4stars = 1e-04,
  pvnshide = T, pvtipl = 0.01, linetype = "n", linecolor = "red",
  linesize = 1, theme = "bw", legend = "top", size = NULL,
  yscale = "none", letleft = T, nletbarmax = 3, font.x = NULL,
  xangle = NULL, yangle = NULL, orientation = "vertical",
  dotsize = NULL, dotshape = NULL, dotcolor = NULL, posd = NULL,
  binwfac = 30, fontfamily = "sans", ymin = NA, ymax = NA,
  fontmain = c(14, "bold", "black"), fontxname = c(12, "plain", "black"),
  fontyname = c(12, "plain", "black"), fontxticks = c(12, "plain",
  "black"), fontyticks = c(12, "plain", "black"), axiscolor = "black",
  tickcolor = "black", axisthick = 0.5, tickthick = 0.5,
  ticklength = 1, xticks.by = NULL, yticks.by = NULL,
  titlejust = "center", legheadsize = 12, legtextsize = 10,
  ftype = NULL, fname = NULL, fscale = NULL, fwidth = NULL,
  fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{ds}{variable or data frame to be summarized (required)}

\item{group1}{optional categorical factor to summarize for each level.}

\item{group2}{optional second factor to summarize for each level.}

\item{doAll}{=T (default) for "All" column; F for no; "I" in quotes for All first.}

\item{brief}{=F (default) to get full summary; T to get just mean&SD or median&IQR.}

\item{pnorm}{=value (default 0.05) to use median&IQR when brief=T if any norm? < value.}

\item{sdsamerow}{=T (default) for SD/IQR on same row when brief=T; F for separate row.}

\item{compare}{=F (default) not to compare levels of groupvar; T to compare.}

\item{minimal}{= FALSE (default), T or TRUE to suppress all output (for cuomics)}

\item{docorr}{=F (default) to not correlate all continuous vars; T to do it.}

\item{plot}{="box" (default) for boxplots; or "bar" "violin" "rod" "no"}

\item{ebars}{=NULL (default)/1/2/3/4 (1-4 for SD/SE/CL/IQR; 4 for median)}

\item{dots}{=1 (default), 0 to not display data on graph}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{barcolor}{for outline color, barfill for fill color. Use fill="grey" ="black" etc for single color.}

\item{barfill}{="lancet" (default) for colors by group ("lancet" for 2 factors, other journal options "aaas", "jco", "uchicago", "npg"}

\item{depname}{/g1name/g2name/title=NULL to override names of dependent/Factor1/Factor2/title}

\item{g1order}{=NULL (default) to reorder bars (1st Factor)}

\item{g2order}{=NULL (default) to reorder bars (2nd Factor)}

\item{caption}{=NULL (default)/"yes"/"caption text" to get caption ("yes" to list n's)}

\item{xlab}{="label" for different x-axis label than name of second argument.}

\item{ylab}{="label" for different y-axis label than name of first argument.}

\item{chariqr}{="," (default) for IQR separator}

\item{charamp}{="&" (default)/x for character to separate group1&2 levels ("-" etc)}

\item{nlevmax}{=9 (default) for maximum number of levels for char-variable}

\item{pnosig}{=0.2 (default)/x for threshold to show p-values even if nonsigificant}

\item{psignif}{=0.05 (default)/x for threshold to significance}

\item{p2stars}{=0.01 (default)/x for threshold to two stars}

\item{p3stars}{=0.001 (default)/x for threshold to three stars}

\item{p4stars}{=0.0001 (default)/x for threshold to four stars}

\item{linetype}{="n" (default)/x for no connecting lines ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linecolor}{="black" (default)/x for black lines ("red" etc)}

\item{linesize}{=1 (default)/x for line thickness}

\item{theme}{="bw" (default)/x for white background (or "gray" or "classic" (no grid lines))}

\item{legend}{(default="top"), can be "bottom", "right", "left"}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can be "horizontal" or "reverse"}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{dotshape}{=NULL (default)}

\item{dotcolor}{=NULL (default "white") set to dot color}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=NULL (default) set to values around 30 to fine-tune horiz.space for dots}

\item{fontfamily}{="sans" (default), can be "serif" "mono"}

\item{ymin, }{ymax =NA (default) or value to start/end y-axis}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for not title}

\item{fontxname}{.fontyname,fontxticks,fontyticks = c(12,"plain","black") default, 0 to suppress}

\item{axiscolor}{,tickcolor="black" (default)/x for axis/tick color}

\item{axisthick}{,tickthick=0.5 (default)/x for axis/tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{,yticks.by =NULL (default)/s for x/y tick spacing by s}

\item{titlejust}{="center" (default) or "left" or "right"}

\item{legheadsize}{=12 (default) for the font size of legend heading}

\item{legtextsize}{=10 (default) for the font size of legend text}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}

\item{pvpairs="std"}{(default)/"all"/i/c(ij,ik,...) to show std or all or vs.i or pval's of i/j, i/k ...}

\item{pvypos=NULL}{(default)/position of pval lines}

\item{pvsi=0.05}{(default)/increment for pvypos from one pval line to next}

\item{pvlab="p"/"*"}{to display numerical or asterisks}

\item{pvprefix="p="}{(default)/"" etc prefix to p-values}

\item{pvsize=3.5}{(default)/size of p-values}

\item{chpvref="ref"}{(default) char above i-th group bar when pvpairs=i}

\item{pvspill=F/T}{to not allow p-values to spill outside range or to allow}

\item{pvnshide=T}{(default)/F to hide NS p-values or not}

\item{pvtipl=0.01}{(default)/length of p-value line tips}
}
\value{
returns summary dataframe (if one variable, like table0)
}
\description{
Table 1 with optional box or violin plots
}
\examples{
attach(NEJM)
cutable1(tgpre, Diet)
detach(NEJM)
cutable1(NEJM, plot="violin")  # single table all variables, violin plots
attach(NEJM)
cutable1(NEJM, Diet, plot="no")  # separate tables all variables, no plots
cutable1(NEJM, Diet, brief=TRUE)  # single table for all variables by Diet columns
cutable1(NEJM, Diet, brief=TRUE, sdsamerow=FALSE)  # to put SD/IQR on separate line
cutable1(NEJM, Diet, brief=TRUE, pnorm=0)  # to get only mean&SD
cutable1(NEJM, Diet, brief=TRUE, pnorm=1)  # to get only median&IQR
detach(NEJM)
}
