% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpubr-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mean_se}
\alias{mean_ci}
\alias{median_q1q3}
\alias{mean_se_}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{mean_se}}}

  \item{ggpubr}{\code{\link[ggpubr:add_summary]{mean_ci}}, \code{\link[ggpubr:add_summary]{mean_se_}}, \code{\link[ggpubr:add_summary]{median_q1q3}}}
}}

