% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volc}
\alias{volc}
\title{Volcano Plot Dataset}
\format{
A data frame with 918 rows and 3 variables:
\describe{
  \item{varnam}{Character: variable/protein name}
  \item{fold}{Numeric: fold change}
  \item{pval}{Numeric: p-value}
}
}
\usage{
volc
}
\description{
Pre-computed fold-change and p-value data for 918 variables,
used to generate volcano plots with \code{cuvolc}.
}
\keyword{datasets}
