% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_rep1fac.R
\name{cu_rep1fac}
\alias{cu_rep1fac}
\title{Does repeated measures categorical data analysis (long-format - multiple lines for each subject)}
\usage{
cu_rep1fac(depvar, groupvar, Subject, scale = "frequency", ebars = 1,
  dots = 0, barcolor = "black", barfill = "colors", plot = "bar",
  order = NULL, psigcld = 0, conf.int = 0.95, depname = NULL,
  g1name = NULL, title = NULL, legend = "top", suff = NULL,
  pvpairs = "all", pvypos = NULL, pvsi = 0.05, pvlab = "p",
  pvprefix = "p=", chpvref = "ref", pvcuts = c(0.2, 0.05, 0.01, 0.001),
  pvnshide = T, pvtipl = 0.01, caption = NULL, linetype = "n",
  linecolor = "black", linesize = 1, theme = "bw", size = NULL,
  width = NULL, yscale = "none", xangle = NULL, yangle = NULL,
  orientation = "vertical", posd = NULL, binwfac = 30, dotsize = NULL,
  dotshape = NULL, dotcolor = NULL, ftype = NULL, fname = NULL,
  fscale = NULL, fwidth = NULL, fheight = NULL, dpi = 300,
  remove = NULL)
}
\arguments{
\item{depvar}{, groupvar, Subject  (3 required)}

\item{scale}{="frequency" (default) or "percent"}

\item{ebars}{=1 (default)/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR)}

\item{dots}{=0 (default), 1 to display data on graph}

\item{barcolor}{for outline color, barfill for fill color. Use fill="colors" for colors by group.}

\item{barfill}{="colors" (default) for colors by group ("lancet" for 2 factors, other journal options "aaas", "jco", "uchicago", "npg"}

\item{plot}{="bar" (default) for bar graphs; "box" "violin" "rod" "no"}

\item{order}{=NULL (default)/c("...") to reorder bars}

\item{psigcld}{=0 (no letters)/x for CLD letters on bars (any shared letter means P>psigcld, 0 for no letters)}

\item{conf.int}{=0.95 (default)/x/0 for confidence interval width of contrast estimates (0 for none)}

\item{depname}{/g1name/title to override names of depvar, groupvar, title}

\item{legend}{="top" (default), "bottom", "right", "left" to locate legend}

\item{linetype}{="n" (default)/x for no connecting lines ("dashed" "dotted")}

\item{linecolor}{="black" (default)/x for black lines ("red" etc)}

\item{linesize}{=1 (default)/x for line thickness}

\item{theme}{="bw" (default)/x for white background (or "gray" or "classic" (no grid lines))}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{xangle}{/yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can change to "horizontal"}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=NULL (default 30) set to fraction of range within which points will be binned}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{dotcolor}{=NULL (default "white") set to dot color}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns nothing
}
\description{
Does repeated measures categorical data analysis (long-format - multiple lines for each subject)
}
\examples{
# Internal function, not exported; called via curepmeas()
\dontrun{
cu_rep1way(TG, Diet, ID)
}
}
