% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_rep1way.R
\name{cu_rep1way}
\alias{cu_rep1way}
\title{Does repeated measures one-way anova and post-hoc tests (long-format - multiple lines for each subject)}
\usage{
cu_rep1way(ddep, dgp, dsub, depvar, groupvar, Subject, itrans, minimal = F,
  ebars = 0, dots = 0, barcolor = "black", barfill = "colors",
  plot = "bar", order = NULL, psigcld = 0, conf.int = 0.95,
  ddepn = NULL, pnorm = 0.05, depname = NULL, g1name = NULL,
  title = NULL, legend = "top", suff = NULL, pvpairs = "std",
  pvypos = NULL, pvstinc = 0.05, pvlab = "p", pvprefix = "p=",
  pvsize = NULL, chpvref = "ref", pvspill = F, pnosig = 0.2,
  psignif = 0.05, p2stars = 0.01, p3stars = 0.001, p4stars = 1e-04,
  pvnshide = T, pvtipl = 0.01, refmean = NA, caption = NULL,
  linetype = "n", linecolor = "black", linesize = 1, theme = "bw",
  size = NULL, width = NULL, yscale = "none", xangle = NULL,
  yangle = NULL, orientation = "vertical", posd = NULL, binwfac = 30,
  dotsize = NULL, dotshape = NULL, dotcolor = NULL, fontmain = NULL,
  fontxname = NULL, fontyname = NULL, fontxticks = NULL,
  fontyticks = NULL, axiscolor = NULL, tickcolor = NULL,
  axisthick = NULL, tickthick = NULL, ticklength = NULL,
  xticks.by = NULL, yticks.by = NULL, ftype = NULL, fname = NULL,
  fscale = NULL, fwidth = NULL, fheight = NULL, dpi = 300,
  remove = NULL)
}
\arguments{
\item{ddep}{required: dependent variable values}

\item{dgp}{required: grouping factor values}

\item{dsub}{required: subject identifier values}

\item{depvar}{char-string name of dependent variable}

\item{groupvar}{char-string name of grouping factor}

\item{Subject}{char-string name of subject identifier}

\item{itrans}{type of transformation if any}

\item{minimal}{= FALSE (default), T or TRUE to suppress all output (for cuomics)}

\item{ebars}{=0 (default to be 1 or 4)/1/2/3 (post-hoc t, SD/SE/CL) or 4 (nonparametric, IQR) or -N (nonparm if any norm fail)}

\item{dots}{=0 (default), 1 to display data on graph}

\item{barcolor}{="black" (default) for bar outline color}

\item{barfill}{="colors" (default) for fill color. "colors" for colors by group ("lancet" for 2 factors, other journal options "aaas", "jco", "uchicago", "npg"}

\item{plot}{="bar" (default) for bar graphs; "box" "violin" "rod" "no"}

\item{order}{=NULL (default)/c("...") to reorder bars}

\item{psigcld}{=0 (no letters)/x for CLD letters on bars (any shared letter means P>psigcld, 0 for no letters)}

\item{conf.int}{=0.95 (default)/x/0 for confidence interval width of contrast estimates (0 for none)}

\item{ddepn}{=NULL (default) for name of dependent variable difference}

\item{pnorm}{=0.05 (default) threshold for normality test}

\item{depname}{=NULL to override name of depvar}

\item{g1name}{=NULL to override name of groupvar}

\item{title}{=NULL to override plot title}

\item{legend}{="top" (default), "bottom", "right", "left" to locate legend}

\item{suff}{=NULL (default) set to desired suffix}

\item{pvpairs}{="std" (default)/"all"/i/c(ij,ik,...) to show std or all or vs.i or pval's of i/j, i/k ...}

\item{pvypos}{=NULL (default)/position of pval lines}

\item{pvstinc}{=0.05 (default)/increment for pvypos from one pval line to next}

\item{pvlab}{="p" (default)/"*" to display numerical or asterisks}

\item{pvprefix}{="p=" (default)/"" etc prefix to p-values}

\item{pvsize}{=NULL (default 7, 3.5)/size of p-values or **}

\item{chpvref}{="ref" (default) char above i-th group bar when pvpairs=i}

\item{pvspill}{=F (default)/T to not allow p-values to spill outside range or to allow}

\item{pnosig}{=0.2 (default)/x for threshold to show p-values even if nonsigificant}

\item{psignif}{=0.05 (default)/x for threshold to significance}

\item{p2stars}{=0.01 (default)/x for threshold to two stars}

\item{p3stars}{=0.001 (default)/x for threshold to three stars}

\item{p4stars}{=0.0001 (default)/x for threshold to four stars}

\item{pvnshide}{=T (default)/F to hide NS p-values or not}

\item{pvtipl}{=0.01 (default)/length of p-value line tips}

\item{refmean}{=NA (default)/ref.val to compare each group mean with}

\item{caption}{=NULL (default)/"yes"/"caption text" to get caption ("yes" to list n's)}

\item{linetype}{="n" (default)/x for no connecting lines ("dashed" "dotted")}

\item{linecolor}{="black" (default)/x for black lines ("red" etc)}

\item{linesize}{=1 (default)/x for line thickness}

\item{theme}{="bw" (default)/x for white background (or "gray" or "classic" (no grid lines))}

\item{size}{numeric value (e.g. size=1), to change size of points and outlines}

\item{width}{numeric value between 0 and 1 specifying box width}

\item{yscale}{(default="none"), can be "log2", "log10", "sqrt"}

\item{xangle}{=NULL for x-axis value angle: 0 (default) horizontal, 90 vertical, or any value between}

\item{yangle}{=NULL for y-axis value angle: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can change to "horizontal"}

\item{posd}{=NULL (default) set to values around 0.9 to fine-tune group2 bar spacing}

\item{binwfac}{=NULL (default 30) set to fraction of range within which points will be binned}

\item{dotsize}{=NULL (default 1) set to fraction of binwidth for dot size}

\item{dotshape}{=NULL (default)}

\item{dotcolor}{=NULL (default "white") set to dot color}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for no title}

\item{fontxname}{=c(12,"plain","black") (default) for x-axis name font, 0 to suppress}

\item{fontyname}{=c(12,"plain","black") (default) for y-axis name font, 0 to suppress}

\item{fontxticks}{=c(12,"plain","black") (default) for x-axis tick label font, 0 to suppress}

\item{fontyticks}{=c(12,"plain","black") (default) for y-axis tick label font, 0 to suppress}

\item{axiscolor}{="black" (default)/x for axis color}

\item{tickcolor}{="black" (default)/x for tick color}

\item{axisthick}{=0.5 (default)/x for axis thickness}

\item{tickthick}{=0.5 (default)/x for tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{=NULL (default)/s for x tick spacing by s}

\item{yticks.by}{=NULL (default)/s for y tick spacing by s}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{=NULL(default) or set to numerical value for file scale}

\item{fwidth}{=NULL(default) or set to file width}

\item{fheight}{=NULL(default) or set to file height}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
one line of summary means/SDs and p-values
}
\description{
Does repeated measures one-way anova and post-hoc tests (long-format - multiple lines for each subject)
}
\examples{
# Internal function, not exported; called via curepmeas()
\dontrun{
cu_rep1way(TG, Diet, ID)
}
}
