% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cufunctions-package.R
\docType{package}
\name{cufunctions-package}
\alias{cufunctions}
\alias{cufunctions-package}
\title{cufunctions: Facilitate Stats with One-line Invokes Instead of Many R Commands}
\description{
Look at help for each function for functionality: cu1way, cu2way, cucov1way, cucov2way, cucox, cukm, culinreg, culogist, cuomics, cupairwise.t, cupartialF, curead, curepmeas, curoc, cuscatter, cutable1, cuvolcano
}
\author{
\strong{Maintainer}: Rajasekhar Ramakrishnan \email{rr6@caa.columbia.edu}

Authors:
\itemize{
  \item Steve Holleran \email{sfh2@columbia.edu}
}

Other contributors:
\itemize{
  \item Janak Ramakrishnan \email{janak@janak.org} (test infrastructure and CRAN readiness) [contributor]
}

}
\keyword{internal}
