% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuomics.R
\name{cuomics}
\alias{cuomics}
\title{Analyzes multiple outcomes by specified function name}
\usage{
cuomics(funcname, dsgiven, dnamf, dnaml, fnom, ...)
}
\arguments{
\item{dsgiven}{,dnamf,dnaml required: dataset, char-strings of names of first&last variable being analyzed}

\item{fnom}{= string as prefix for output file cuomics.csv}

\item{...}{remaining arguments for chosen function}

\item{funcname:}{"1way","2way","cov1way","cov2way","scatter","repmeasl","repmeasw"}
}
\value{
nothing
}
\description{
Analyzes multiple outcomes by specified function name
}
\examples{
d <- tempdir()  # change to "." to write to current directory
attach(NEJM)
cuomics("1way", NEJM, "tcstudy", "hcchange", file.path(d, "tchc1w"), Diet)
cuomics("2way", NEJM, "tcstudy", "hcchange", file.path(d, "tchc2w"), Diet, sex)
# TODO(sekhar): fix these examples, they hit "stop" case below
\dontrun{
cuomics("cov1way", NEJM, "tcstudy", "hcchange", file.path(d, "tchccov1"), tcpre, Diet)
cuomics("cov2way", NEJM, "tcstudy", "hcchange", file.path(d, "tchccov2"), tcpre, Diet, sex)
cuomics("scatter", NEJM, "tcstudy", "hcchange", file.path(d, "tchcscat"), tcpre)
}
detach(NEJM)
# TODO: requires fix-cuomics-normvisit branch to be merged
\dontrun{
attach(delta)
cuomics("repmeasw", delta, "TC", "TG", file.path(d, "tctgrep"), "Diet", "sex")
detach(delta)
}
}
