% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curepmany.R
\name{curepmany}
\alias{curepmany}
\title{Does repeated measures one- or two-way anova and post-hoc tests (format wide or long - 1/many lines for each subject)}
\usage{
curepmany(dsgiven, dnamf, dnaml, repnam, fac2 = NULL, idnam = NULL,
  minimal = T, interact = TRUE, partialF = TRUE, cov = NULL,
  ytrans = "none", maxlines = 0, mainx = 1, sumdiff = 1,
  scale = "frequency", scatter = "yes", ebars = 1, dots = 0,
  barcolor = "black", barfill = NULL, plot = "bar", ordinal = NULL,
  g1order = NULL, g2order = NULL, difforder = NULL, depname = NULL,
  g1name = NULL, g2name = NULL, title = NULL, chariqr = ":",
  psigcld = 0, conf.int = 0.95, ddepn = NULL, shape = 16,
  caption = NULL, pvpairs = "std", pvypos = NULL, pvstinc = 0.05,
  pvlab = "p", pvprefix = "p=", pvsize = NULL, chpvref = "ref",
  pvspill = F, pnosig = 0.2, psignif = 0.05, p2stars = 0.01,
  p3stars = 0.001, p4stars = 1e-04, pvnshide = T, pvtipl = 0.01,
  normvisit = 0, normway = "fold", needvar = 1, linetype = "n",
  linecolor = "black", linesize = 1, theme = "bw", legend = "top",
  size = NULL, width = NULL, yscale = "none", xangle = NULL,
  yangle = NULL, orientation = "vertical", posd = NULL, binwfac = 30,
  dotsize = NULL, dotshape = NULL, dotcolor = NULL, fontmain = NULL,
  fontxname = NULL, fontyname = NULL, fontxticks = NULL,
  fontyticks = NULL, axiscolor = NULL, tickcolor = NULL,
  axisthick = NULL, tickthick = NULL, ticklength = NULL,
  xticks.by = NULL, yticks.by = NULL, dodredge = T, usemod = 1,
  nmodshow = 16, nverscat = 2, nhorscat = 2, ifdeb = F, m.min = 1,
  m.max = 99, fixed = NULL, subset = as.expression(1), ftype = NULL,
  fname = NULL, fscale = NULL, fwidth = NULL, fheight = NULL,
  dpi = 300, remove = NULL)
}
\arguments{
\item{dsgiven}{,dnamf,dnaml,repnam required: dataset, char-strings of names of first&last variable being analyzed, repeated grouping factor}

\item{fac2}{char-string of name of 2nd, non-repeated grouping factor (default NULL)}

\item{idnam}{(to indicate long format) =char-string of name of ID/Subject variable (default NULL)}

\item{minimal}{= TRUE (default), F or FALSE to get all console output and graphs}
}
\value{
nothing
}
\description{
Does repeated measures one- or two-way anova and post-hoc tests (format wide or long - 1/many lines for each subject)
}
\examples{
\dontrun{
curepmeas(delta,"TC","TG","Diet")  # one line per subject; TG.xx TG.yy etc as names of TG data
curepmeas(delta,"TC","TG","Diet", "sex")  # with sex as second factor, Diet*sex interaction in model
curepmeas(delta,"TC","TG","Diet", "sex", interact=F)  # no Diet*sex interaction
curepmeas(delta,"TC","TG","Diet", "sex", cov="age+sex*age")  # with covariate adjustment
}
}
