% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curoc.R
\name{curoc}
\alias{curoc}
\title{For logistic-regression, does ROC curve(s), AUC, and summary table with best accuracy}
\usage{
curoc(LRobj, depvar, twolev = TRUE, xs = NULL, namedep = NULL,
  logitlog = "logit", printfit = F, emf = F,
  xlabroc = "1 - Specificity", ylabroc = "Sensitivity",
  label.ordering = NULL, color = "red", veccolor = c("red", "blue",
  "green", "black"))
}
\arguments{
\item{LRobj}{logistic-regression object (glm or polr) (required)}

\item{depvar}{outcome variable (required)}

\item{twolev}{=T (default) for glm object; =F for polr object (>2 levels)}

\item{xs}{list of probability cut points for 2x2 tables (besides 0.5)}

\item{namedep}{outcome variable name labeling graph}

\item{logitlog}{="logit" (default) to model odds ("logit") or risk ("log")}

\item{emf}{=F(default)/T/name (to create culogist.emf or name.emf needed on Mac)}

\item{xlabroc}{="x axis label" (default "1 - Specificity")}

\item{ylabroc}{="y axis label" (default "Sensitivity")}

\item{label.ordering}{= c(two labels) to override alphabetical order}

\item{color}{="red" (default) for single ROC curve color (glm object)}

\item{veccolor}{=c("red","blue","green","black") (default) for multiple curves (polr object)}
}
\value{
returns nothing
}
\description{
For logistic-regression, does ROC curve(s), AUC, and summary table with best accuracy
}
\examples{
logfit <- glm(MetSyn ~ TG + HDL, data = Met, family = binomial)
curoc(logfit, Met$MetSyn)
curoc(logfit, Met$MetSyn, xs=c(0.3, 0.7))  # AUC with 3 summary tables
polrfit <- MASS::polr(as.factor(WTCAT) ~ TG + HDL, data = Met)
curoc(polrfit, Met$WTCAT, twolev=FALSE)  # multi-level via polr
}
