% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuscatter.R
\name{cuscatter}
\alias{cuscatter}
\title{scatterplot of yvar vs xvar}
\usage{
cuscatter(yvar, xvar, doline = T, showr2eqn = "both", minimal = F,
  yname = NULL, xname = NULL, title = NULL, caption = "",
  xangle = NULL, yangle = NULL, orientation = "vertical", shape = 1,
  dotcolor = "black", dotsize = 2, linetype = "solid", linesize = 1,
  linecolor = "red", theme = "bw", xscale = "none", yscale = "none",
  fontfamily = "sans", r2x = 1, r2y = 0.95, r2size = 4,
  r2color = "blue", r2font = "plain", xmin = NA, ymin = NA,
  xmax = NA, ymax = NA, fontmain = c(14, "bold", "black"),
  fontxname = c(12, "plain", "black"), fontyname = c(12, "plain", "black"),
  fontxticks = c(12, "plain", "black"), fontyticks = c(12, "plain",
  "black"), axiscolor = "black", tickcolor = "black", axisthick = 0.5,
  tickthick = 0.5, ticklength = 1, xticks.by = NULL, yticks.by = NULL,
  titlejust = "center", legheadsize = 12, legtextsize = 10,
  ftype = NULL, fname = NULL, fscale = NULL, fwidth = NULL,
  fheight = NULL, dpi = 300, remove = NULL)
}
\arguments{
\item{yvar}{,xvar required}

\item{doline}{=T (default) to show the regression line, F to not show}

\item{showr2eqn}{="both" (default) to show R2 and eqn,"r2" for just R2, "eqn" for just eqn, "no" for neither}

\item{minimal}{=F (default)/T for minimal console output and no graph}

\item{yname}{,xname,title ="xyz" to override names of yvar, xvar, title}

\item{caption}{='' (default) or set to string to show at bottom right}

\item{xangle}{,yangle for axis value angles: 0 (default) horizontal, 90 vertical, or any value between}

\item{orientation}{(default="vertical"), can be "horizontal" or "reverse"}

\item{shape}{=1 (default) (see http://www.sthda.com/english/wiki/ggplot2-point-shapes)}

\item{dotcolor}{="black" (default) for point symbol color (="red" ="blue" etc)}

\item{dotsize}{=2 (default) or x to set size of point symbols}

\item{linetype}{="solid" (default)/x for solid line ("solid" "dashed" "dotted" "blank" "longdash" "dotdash" "twodash")}

\item{linesize}{=1 (default)/x for line thickness}

\item{linecolor}{="red" (default)/x for line color}

\item{theme}{="bw" (default)/x for white background ("classic" (no grid lines),"linedraw" "gray" "minimal" "void")}

\item{xscale}{,yscale ="none" (default), can be "log2", "log10", "sqrt"}

\item{fontfamily}{="sans" (default), can be "serif" "mono"}

\item{r2x}{=1 (default)/x for x-coordinate as fraction of axis of R2/eqn}

\item{r2y}{=0.95 (default)/x for y-coordinate as fraction of axis of R2/eqn}

\item{r2size}{=4 (default)/x for size of R2/eqn}

\item{r2color}{="blue" (default)/x for R2/eqn color}

\item{r2font}{="plain" (default)/x for R2/eqn font ("bold","italic","bold.italic")}

\item{xmin}{,xmax,ymin,ymax =NA (default) or value to start/end x/y-axis}

\item{fontmain}{=c(14,"bold","black") default, change for title, 0 for not title}

\item{fontxname}{.fontyname,fontxticks,fontyticks = c(12,"plain","black") default, 0 to suppress}

\item{axiscolor}{,tickcolor="black" (default)/x for axis/tick color}

\item{axisthick}{,tickthick=0.5 (default)/x for axis/tick thickness}

\item{ticklength}{=1 (default)/x for tick length in mm}

\item{xticks.by}{,yticks.by =NULL (default)/s for x/y tick spacing by s}

\item{titlejust}{="center" (default) or "left" or "right"}

\item{ftype}{=NULL(default)/eps/pdf/jpg/jpeg/tiff/png/emf (for hires file or name.emf for Mac)}

\item{fname}{=NULL(default) or set to prefix for "funcname.ftype"}

\item{fscale}{,fwidth,fheight =NULL(default) or set to numerical value}

\item{dpi}{=300 (default) or set to desired resolution in dpi in file}

\item{remove}{choose from =c("xlab","ylab","x.text","y.text","x.ticks","y.ticks","grid","x.grid","y.grid","axis","x.axis","y.axis")}
}
\value{
returns list with plot object, slope, intercept, pval, R^2
}
\description{
scatterplot of yvar vs xvar
}
\examples{
attach(NEJM)
cuscatter(tcstudy, tcpre)
detach(NEJM)
}
