% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvolcano.R
\name{cuvolcano}
\alias{cuvolcano}
\title{Volcano plot}
\usage{
cuvolcano(dsomics, gpairs, xdash = 0.5, psig = 0.05, psigmin = 0.2,
  pfactor = 1, foldmin = NULL, foldmax = NULL, ylim = 5,
  titlend = "Volcano Plot", ftype = NULL, fname = NULL, fscale = NULL,
  fwidth = NULL, fheight = NULL, dpi = 300, remove = NULL,
  suff = NULL)
}
\arguments{
\item{dsomics}{data frame from cuomics}

\item{gpairs}{list of group pairs to be compared}

\item{xdash}{=0.5 (default) to locate dashed vertical lines}

\item{psig}{=0.05 (default) for p-value threshold of interest}

\item{psigmin}{=0.2 (default) for upper limit of p-values to display}

\item{pfactor}{=1 (default) for factor to divide psig by for multiple comparison correction}

\item{foldmin}{=NULL (default) for left end of log2(FoldChange) (usually set to -2)}

\item{foldmax}{=NULL (default) for right end of log2(FoldChange) (usually set to 2)}

\item{ylim}{=5 (default) for upper limit of graph (-log10(pval))}
}
\description{
Volcano plot
}
\examples{
cuvolcano(volcomic, c("lean", "obese"))
}
